+++ src/algebra/scalar_multiplication/multiexp.hpp
@@ -42,7 +42,7 @@ T multi_exp(typename std::vector<T>::const_iterator vec_start,
             typename std::vector<T>::const_iterator vec_end,
             typename std::vector<FieldT>::const_iterator scalar_start,
             typename std::vector<FieldT>::const_iterator scalar_end,
-            const size_t chunks,
+            const uint64_t chunks,
             const bool use_multiexp=false);
 
 
@@ -54,7 +54,7 @@ T multi_exp_with_mixed_addition(typename std::vector<T>::const_iterator vec_star
                                   typename std::vector<T>::const_iterator vec_end,
                                   typename std::vector<FieldT>::const_iterator scalar_start,
                                   typename std::vector<FieldT>::const_iterator scalar_end,
-                                  const size_t chunks,
+                                  const uint64_t chunks,
                                   const bool use_multiexp);
 
 /**
@@ -67,31 +67,31 @@ using window_table = std::vector<std::vector<T> >;
  * Compute window size for the given number of scalars.
  */
 template<typename T>
-size_t get_exp_window_size(const size_t num_scalars);
+uint64_t get_exp_window_size(const uint64_t num_scalars);
 
 /**
  * Compute table of window sizes.
  */
 template<typename T>
-window_table<T> get_window_table(const size_t scalar_size,
-                                 const size_t window,
+window_table<T> get_window_table(const uint64_t scalar_size,
+                                 const uint64_t window,
                                  const T &g);
 
 template<typename T, typename FieldT>
-T windowed_exp(const size_t scalar_size,
-               const size_t window,
+T windowed_exp(const uint64_t scalar_size,
+               const uint64_t window,
                const window_table<T> &powers_of_g,
                const FieldT &pow);
 
 template<typename T, typename FieldT>
-std::vector<T> batch_exp(const size_t scalar_size,
-                         const size_t window,
+std::vector<T> batch_exp(const uint64_t scalar_size,
+                         const uint64_t window,
                          const window_table<T> &table,
                          const std::vector<FieldT> &v);
 
 template<typename T, typename FieldT>
-std::vector<T> batch_exp_with_coeff(const size_t scalar_size,
-                                    const size_t window,
+std::vector<T> batch_exp_with_coeff(const uint64_t scalar_size,
+                                    const uint64_t window,
                                     const window_table<T> &table,
                                     const FieldT &coeff,
                                     const std::vector<FieldT> &v);
