+++ src/gadgetlib1/gadgets/hashes/hash_io.hpp
@@ -15,15 +15,15 @@ namespace libsnark {
 template<typename FieldT>
 class digest_variable : public gadget<FieldT> {
 public:
-    size_t digest_size;
+    uint64_t digest_size;
     pb_variable_array<FieldT> bits;
 
     digest_variable<FieldT>(protoboard<FieldT> &pb,
-                            const size_t digest_size,
+                            const uint64_t digest_size,
                             const std::string &annotation_prefix);
 
     digest_variable<FieldT>(protoboard<FieldT> &pb,
-                            const size_t digest_size,
+                            const uint64_t digest_size,
                             const pb_variable_array<FieldT> &partial_bits,
                             const pb_variable<FieldT> &padding,
                             const std::string &annotation_prefix);
@@ -36,11 +36,11 @@ public:
 template<typename FieldT>
 class block_variable : public gadget<FieldT> {
 public:
-    size_t block_size;
+    uint64_t block_size;
     pb_variable_array<FieldT> bits;
 
     block_variable(protoboard<FieldT> &pb,
-                   const size_t block_size,
+                   const uint64_t block_size,
                    const std::string &annotation_prefix);
 
     block_variable(protoboard<FieldT> &pb,
