+++ src/gadgetlib1/gadgets/hashes/hash_io.tcc
@@ -11,7 +11,7 @@ namespace libsnark {
 
 template<typename FieldT>
 digest_variable<FieldT>::digest_variable(protoboard<FieldT> &pb,
-                                         const size_t digest_size,
+                                         const uint64_t digest_size,
                                          const std::string &annotation_prefix) :
     gadget<FieldT>(pb, annotation_prefix), digest_size(digest_size)
 {
@@ -20,7 +20,7 @@ digest_variable<FieldT>::digest_variable(protoboard<FieldT> &pb,
 
 template<typename FieldT>
 digest_variable<FieldT>::digest_variable(protoboard<FieldT> &pb,
-                                         const size_t digest_size,
+                                         const uint64_t digest_size,
                                          const pb_variable_array<FieldT> &partial_bits,
                                          const pb_variable<FieldT> &padding,
                                          const std::string &annotation_prefix) :
@@ -37,7 +37,7 @@ digest_variable<FieldT>::digest_variable(protoboard<FieldT> &pb,
 template<typename FieldT>
 void digest_variable<FieldT>::generate_r1cs_constraints()
 {
-    for (size_t i = 0; i < digest_size; ++i)
+    for (uint64_t i = 0; i < digest_size; ++i)
     {
         generate_boolean_r1cs_constraint<FieldT>(this->pb, bits[i], FMT(this->annotation_prefix, " bits_%zu", i));
     }
@@ -57,7 +57,7 @@ bit_vector digest_variable<FieldT>::get_digest() const
 
 template<typename FieldT>
 block_variable<FieldT>::block_variable(protoboard<FieldT> &pb,
-                                       const size_t block_size,
+                                       const uint64_t block_size,
                                        const std::string &annotation_prefix) :
     gadget<FieldT>(pb, annotation_prefix), block_size(block_size)
 {
